requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-multipart>=0.0.6
numpy>=1.24.0
pydantic>=2.5.0